% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FALightning.R
\name{e_step}
\alias{e_step}
\title{EM Algorithm for Factor Analsysis: E-step}
\usage{
e_step(x, lambda, phi)
}
\arguments{
\item{x}{Sample matrix (dimension N by P')}

\item{lambda}{Factor loading matrix (dimension P by P')}

\item{phi}{Vector of noise variance}
}
\value{
A list containing:
\describe{
\item{ez}{Expected latent scores from expected_scores()}
\item{ezz}{Expected covariance of the latent factors from expected_cov()}
\item{xez}{Covariance between observed and latent factors}
}
}
\description{
EM Algorithm for Factor Analsysis: E-step
}
\note{
For Internal Use
}
