% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FALightning.R
\name{expected_cov}
\alias{expected_cov}
\title{Expected Covariance of the Latent Factors}
\usage{
expected_cov(ez, beta, lambda)
}
\arguments{
\item{ez}{Expected latent scores from expected_scores()}

\item{beta}{Coefficients from get_beta()}

\item{lambda}{Factor loading matrix (dimension P by P')}
}
\value{
The covariance of the expected latent scores
}
\description{
Expected Covariance of the Latent Factors
}
\note{
For Internal Use
}
