% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FALightning.R
\name{expected_scores}
\alias{expected_scores}
\title{Expected Latent Scores}
\usage{
expected_scores(x, beta)
}
\arguments{
\item{x}{Sample matrix (dimension N by P')}

\item{beta}{Coefficients from get_beta()}
}
\value{
Expected latent scores conditioned on observed data
}
\description{
Expected Latent Scores
}
\note{
For Internal Use
}
