% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FALightning.R
\name{factor_analyzer}
\alias{factor_analyzer}
\title{EM Algorithm for Factor Analsysis}
\usage{
factor_analyzer(
  x,
  n_factors,
  n_iter = 200,
  tol = 1e-04,
  rotation = varimax,
  verbose = F,
  ...
)
}
\arguments{
\item{x}{Sample matrix (dimension N by P)}

\item{n_factors}{An integer: the number of factors}

\item{n_iter}{Number of EM iterations}

\item{tol}{Tolerance for convergence}

\item{rotation}{A function for factor rotation, e.g. varimax or oblimin from the GPArotation package}

\item{verbose}{Whether to display EM updates}

\item{...}{Other parameters passed to the "rotation" function}
}
\value{
A list containing:
\describe{
  \item{loadings}{If rotation = NULL, the factor loadings estimate. Otherwise, the object returned by the "rotation" function.}
  \item{phi}{The phi vector estimate}
  \item{scores}{The expected factor scores}
  \item{crit}{The log-likelihood criterion per iteration}
  \item{converge_status}{Convergence status: 0 = converged, 1 = tolerance not reached, 2 = likelihood has decreased}
}
}
\description{
EM Algorithm for Factor Analsysis
}
\examples{
set.seed(8)
z = matrix(rnorm(2000), 1000, 2)
lambda_orig = matrix(rnorm(20, sd = 1), nrow = 10, ncol = 2)
x = z \%*\% t(lambda_orig) + matrix(rnorm(10000, sd = sqrt(0.1)), 1000, 10)
fit = factor_analyzer(x, 2, rotation = NULL)
plot(fit$crit, type = "l")
fit$loadings
fit_varimax = factor_analyzer(x, 2)
fit_varimax$loadings
fit_promax = factor_analyzer(x, 2, rotation = promax)
fit_promax$loadings
}
