% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FALightning.R
\name{factor_analyzer_cov}
\alias{factor_analyzer_cov}
\title{EM Algorithm for Factor Analsysis with Covariance Input}
\usage{
factor_analyzer_cov(
  cov_x,
  n_factors,
  covar = NULL,
  n_sample = 10000,
  ini_method = "pca",
  breath_of_lightning = T,
  ...
)
}
\arguments{
\item{cov_x}{Sample covariance/correlation matrix}

\item{n_factors}{An integer: the number of factors}

\item{covar}{Whether input cov_x is a covariance (T) or correlation (F) matrix}

\item{n_sample}{Number of samples to draw from Multivariate Gaussian}

\item{ini_method}{The EM initialization method ("pca" recommended)}

\item{...}{Other parameters passed to the "factor_analyzer" function}

\item{breath_of_lighting}{Whether to use irlba for initialization (recommended)}
}
\value{
A fitted factor analyzer object
}
\description{
Performs Factor Analysis given a covariance/correlation matrix.
}
\note{
The function runs EM on a simulated multivariate Gaussian sample
to obtain approximate maximum likelihood estimates. n_sample, the number of
samples to simulate, should be large (200000 based on our benchmark experiments).
As EM will be performed on this large data set, initializing EM by
pca + breath_of_lightning is recommended.
}
\examples{
\dontrun{
set.seed(8)
z = matrix(rnorm(2000), 1000, 2)
lambda_orig = matrix(rnorm(20, sd = 1), nrow = 10, ncol = 2)
x = z \%*\% t(lambda_orig) + matrix(rnorm(10000, sd = sqrt(0.1)), 1000, 10)

fit = factor_analyzer(x, 2)
varimax(fit$loadings)
cov_x = cov(x)*(nrow(x)-1)/nrow(x)
fit_cov = factor_analyzer_cov(cov_x, 2, covar = T, n_sample = 200000)
varimax(fit_cov$loadings)

x = scale(x)
fit = factor_analyzer(x, 2)
varimax(fit$loadings)
cor_x = cor(x)
fit_cor = factor_analyzer_cov(cor_x, 2, covar = F, n_sample = 200000)
varimax(fit_cor$loadings)
}
}
