% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FALightning.R
\name{fast_get_beta}
\alias{fast_get_beta}
\title{Coefficients for Evaluating the Expected Latent Factors: Fast Calculation}
\usage{
fast_get_beta(lambda, phi)
}
\arguments{
\item{lambda}{Factor loading matrix (dimension P by P')}

\item{phi}{Vector of noise variance}
}
\value{
A set of coefficients for evaluating the expected latent factors
}
\description{
This function merged two older functions get_beta() and fast_inverse().
This allows speed ups as the entire inverse covariance is never necessary
in the EM fitting procedure. Speed up is thus achieved by only
performing the calculations necessary for the coefficients.
}
\note{
For Internal Use
}
