% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FALightning.R
\name{get_beta}
\alias{get_beta}
\title{Coefficients for Evaluating the Expected Latent Factors}
\usage{
get_beta(lambda, inverse)
}
\arguments{
\item{lambda}{Factor loading matrix (dimension P by P')}

\item{inverse}{Inverse covariance matrix}
}
\value{
A set of coefficients for evaluating the expected latent factors
}
\description{
Coefficients for Evaluating the Expected Latent Factors
}
\note{
For Internal Use. Deprecated as of Jan 5, 2023
}
