% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FALightning.R
\name{get_scores}
\alias{get_scores}
\title{Factor Scores}
\usage{
get_scores(fit, newdata)
}
\arguments{
\item{fit}{A fitted object from factor_analyzer()}

\item{newdata}{A new data set for scores computation}
}
\value{
The projected scores of newdata
}
\description{
Factor Scores
}
\examples{
set.seed(8)
z = matrix(rnorm(3000), 1000, 3)
z_2 = matrix(rnorm(3000), 1000, 3)
lambda_orig = matrix(rnorm(30, sd = 1), nrow = 10, ncol = 3)
x = z \%*\% t(lambda_orig) + matrix(rnorm(10000, sd = sqrt(0.1)), 1000, 10)
newdata = z_2 \%*\% t(lambda_orig) + matrix(rnorm(10000, sd = sqrt(0.1)), 1000, 10)
fit = factor_analyzer(x, 2)
scores = get_scores(fit, newdata)
}
