% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FALightning.R
\name{get_xEz}
\alias{get_xEz}
\title{Covariance Between Observed Samples and Latent Factors}
\usage{
get_xEz(x, ez)
}
\arguments{
\item{x}{Sample matrix (dimension N by P)}

\item{ez}{Expected latent scores from expected_scores()}
}
\value{
Covariance between observed and latent scores
}
\description{
Covariance Between Observed Samples and Latent Factors
}
\note{
For Internal Use
}
