% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FALightning.R
\name{loglik}
\alias{loglik}
\title{Log-Likelihood}
\usage{
loglik(x_cov, lambda, phi, e_obj)
}
\arguments{
\item{x_cov}{Sample covariance matrix}

\item{lambda}{Factor loading matrix (dimension P by P')}

\item{phi}{Vector of noise variance}

\item{e_obj}{A list of expected values from the E-Step}
}
\value{
The log-likelihood criterion
}
\description{
Log-Likelihood
}
\note{
For Internal Use
}
