% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FALightning.R
\name{loglik_cov}
\alias{loglik_cov}
\title{Log-Likelihood}
\usage{
loglik_cov(fit, cov_x, n)
}
\arguments{
\item{fit}{A fitted object from factor_analyzer()}

\item{n}{The sample size}
}
\value{
The log-likelihood criterion
}
\description{
Log-Likelihood
}
\note{
For Internal Use
}
