% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FALightning.R
\name{lrt_fa}
\alias{lrt_fa}
\title{Likeilood Ratio Test for Goodness of Fit}
\usage{
lrt_fa(cov_x, n, fit)
}
\arguments{
\item{cov_x}{The sample covariance matrix}

\item{n}{The sample size}

\item{fit}{A fitted object from factor_analyzer()}
}
\value{
A list contain the chi-square statistics (chi_sq), the degree-of-freedom (df), and the p-value.
}
\description{
Likeilood Ratio Test for Goodness of Fit
}
\examples{
set.seed(8)
z = matrix(rnorm(3000), 1000, 3)
lambda_orig = matrix(rnorm(30, sd = 1), nrow = 10, ncol = 3)
x = z \%*\% t(lambda_orig) + matrix(rnorm(10000, sd = sqrt(0.1)), 1000, 10)
s = cov(x)/nrow(x)*(nrow(x) - 1)
p_vec = rep(0, 5)
for (i in seq_along(p_vec)) p_vec[i] = lrt_fa(s, nrow(x), factor_analyzer(x, i))$p_val
p_vec
}
