% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FALightning.R
\name{update_lambda}
\alias{update_lambda}
\title{Lambda Update}
\usage{
update_lambda(xez, ezz)
}
\arguments{
\item{xez}{Covariance between observed and latent factors}

\item{ezz}{Expected covariance of the latent factors from expected_cov()}
}
\value{
The updated lambda matrix
}
\description{
Lambda Update
}
\note{
For Internal Use
}
