% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FALightning.R
\name{e_step}
\alias{e_step}
\title{EM Algorithm for Factor Analsysis: E-step}
\usage{
e_step(x, lambda, phi)
}
\arguments{
\item{x}{Sample matrix (dimension N by P')}

\item{beta}{Coefficients from get_beta()}
}
\value{
A list containing:

ez: Expected latent scores from expected_scores()

ezz: Expected covariance of the latent factors from expected_cov()

xez: Covariance between observed and latent factors
}
\description{
EM Algorithm for Factor Analsysis: E-step
}
\note{
For Internal Use
}
