% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FALightning.R
\name{loglik}
\alias{loglik}
\title{Expected Log-Likelihood}
\usage{
loglik(x, dxx, lambda, phi)
}
\arguments{
\item{x}{Sample matrix (dimension N by P)}

\item{dxx}{Column sums-of-squares of the sample matrix}

\item{lambda}{Factor loading matrix (dimension P by P')}

\item{phi}{Vector of noise variance}
}
\value{
The Expected log-likelihood for EM convergence monitoring
}
\description{
Expected Log-Likelihood
}
\note{
For Internal Use
}
