% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FALightning.R
\name{m_step}
\alias{m_step}
\title{EM Algorithm for Factor Analsysis: M-step}
\usage{
m_step(x, e_obj)
}
\arguments{
\item{x}{Sample matrix (dimension N by P')}

\item{e_obj}{A list of expected values from the E-Step}
}
\value{
A list containing:

lambda: The updated lambda matrix

phi: The updated phi vector
}
\description{
EM Algorithm for Factor Analsysis: M-step
}
\note{
For Internal Use
}
