% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FALightning.R
\name{update_phi}
\alias{update_phi}
\title{Phi Update}
\usage{
update_phi(xx, n, xez, lambda)
}
\arguments{
\item{n}{Sample size}

\item{xez}{Covariance between observed and latent factors}

\item{lambda}{Factor loading matrix (dimension P by P')}

\item{x}{Cross-product matrix of the sample matrix}
}
\value{
The updated phi vector
}
\description{
Phi Update
}
\note{
For Internal Use
}
